refactored files to separate ROS-dependent functions and classes from non-ROS-dependent functions and files
