+++ panther/include/timer.hpp
@@ -2,7 +2,6 @@
 #define TIMER_HPP_
 
 #include <chrono>
-#include <ros/ros.h>
 
 namespace PANTHER_timers
 {
@@ -56,58 +55,6 @@ private:
   high_resolution_clock::time_point start_;
   high_resolution_clock::time_point end_;
 };
-
-class ROSTimer
-{
-public:
-  ROSTimer(bool run = false)
-  {
-    if (run)
-      tic();
-  }
-  void tic()
-  {
-    start_ = ros::Time::now().toSec();
-  }
-  double elapsedSoFarMs() const
-  {
-    return 1000 * (ros::Time::now().toSec() - start_);
-  }
-  template <typename T, typename Traits>
-  friend std::basic_ostream<T, Traits>& operator<<(std::basic_ostream<T, Traits>& out, const ROSTimer& timer)
-  {
-    return out << timer.elapsedSoFarMs();
-  }
-
-private:
-  double start_;
-};
-
-class ROSWallTimer
-{
-public:
-  ROSWallTimer(bool run = false)
-  {
-    if (run)
-      tic();
-  }
-  void tic()
-  {
-    start_ = ros::WallTime::now().toSec();
-  }
-  double elapsedSoFarMs() const
-  {
-    return 1000 * (ros::WallTime::now().toSec() - start_);
-  }
-  template <typename T, typename Traits>
-  friend std::basic_ostream<T, Traits>& operator<<(std::basic_ostream<T, Traits>& out, const ROSWallTimer& timer)
-  {
-    return out << timer.elapsedSoFarMs();
-  }
-
-private:
-  double start_;
-};
-}  // namespace PANTHER_timers
+} // namesapce PANTHER_timers
 
 #endif  // TIMER_HPP_
