+++ panther/include/utils.hpp
@@ -9,22 +9,10 @@
 #ifndef UTILS_HPP
 #define UTILS_HPP
 
-#include <std_msgs/ColorRGBA.h>
-#include <geometry_msgs/Vector3.h>
-#include <geometry_msgs/Point.h>
-#include <visualization_msgs/Marker.h>
-
-#include "tf2_geometry_msgs/tf2_geometry_msgs.h"
-#include "visualization_msgs/Marker.h"
-#include "visualization_msgs/MarkerArray.h"
 #include "panther_types.hpp"
 #include <deque>
-
-#include <panther_msgs/PieceWisePolTraj.h>
-#include <panther_msgs/CoeffPoly.h>
-#include <panther_msgs/Log.h>
-
-#include "ros/ros.h"
+#include <filesystem>
+#include <iterator>
 
 #define RED_NORMAL 1
 #define RED_TRANS 2
@@ -56,17 +44,6 @@
 // #define UNKOWN_AND_OCCUPIED_SPACE 2
 
 // TODO: put this in a namespace
-
-template <typename T>
-bool safeGetParam(ros::NodeHandle& nh, std::string const& param_name, T& param_value)
-{
-  if (!nh.getParam(param_name, param_value))
-  {
-    ROS_ERROR("Failed to find parameter: %s", nh.resolveName(param_name, true).c_str());
-    exit(1);
-  }
-  return true;
-}
 
 // https://stackoverflow.com/questions/27028226/python-linspace-in-c
 template <typename T>
@@ -111,40 +88,11 @@ double getMinTimeDoubleIntegrator3D(const Eigen::Vector3d& p0, const Eigen::Vect
 double getMinTimeDoubleIntegrator3DFromState(mt::state initial_state, mt::state final_state,
                                              const Eigen::Vector3d& v_max, const Eigen::Vector3d& a_max);
 
-panther_msgs::Log log2LogMsg(mt::log log);
-visualization_msgs::MarkerArray pwp2ColoredMarkerArray(mt::PieceWisePol& pwp, double t_init, double t_final,
-                                                       int samples, std::string ns, Eigen::Vector3d& color);
-
-mt::PieceWisePol createPwpFromStaticPosition(const mt::state& current_state);
-
-mt::PieceWisePol pwpMsg2Pwp(const panther_msgs::PieceWisePolTraj& pwp_msg);
-panther_msgs::PieceWisePolTraj pwp2PwpMsg(const mt::PieceWisePol& pwp);
-
-visualization_msgs::Marker edges2Marker(const mt::Edges& edges, std_msgs::ColorRGBA color_marker);
-
-geometry_msgs::Pose identityGeometryMsgsPose();
-
-mt::PieceWisePol composePieceWisePol(const double t, const double dc, mt::PieceWisePol& p1, mt::PieceWisePol& p2);
-
 bool boxIntersectsSphere(Eigen::Vector3d center, double r, Eigen::Vector3d c1, Eigen::Vector3d c2);
 
 void printStateDeque(std::deque<mt::state>& data);
 
-std::vector<std::string> pieceWisePol2String(const mt::PieceWisePol& piecewisepol);
-
 void printStateVector(std::vector<mt::state>& data);
-
-std_msgs::ColorRGBA getColorJet(double v, double vmin, double vmax);
-
-std_msgs::ColorRGBA getColorInterpBetween2Colors(double v, double vmin, double vmax, std_msgs::ColorRGBA min_color,
-                                                 std_msgs::ColorRGBA max_color);
-std_msgs::ColorRGBA color(int id);
-
-void quaternion2Euler(tf2::Quaternion q, double& roll, double& pitch, double& yaw);
-
-void quaternion2Euler(Eigen::Quaterniond q, double& roll, double& pitch, double& yaw);
-
-void quaternion2Euler(geometry_msgs::Quaternion q, double& roll, double& pitch, double& yaw);
 
 void saturate(int& var, const int min, const int max);
 
@@ -152,25 +100,11 @@ void saturate(double& var, const double min, const double max);
 
 void saturate(Eigen::Vector3d& tmp, const Eigen::Vector3d& min, const Eigen::Vector3d& max);
 
-visualization_msgs::Marker getMarkerSphere(double scale, int my_color);
-
 double angleBetVectors(const Eigen::Vector3d& a, const Eigen::Vector3d& b);
 
 void angle_wrap(double& diff);
 
-geometry_msgs::Point pointOrigin();
-
-Eigen::Vector3d vec2eigen(geometry_msgs::Vector3 vector);
-
-geometry_msgs::Vector3 eigen2rosvector(Eigen::Vector3d vector);
-
-geometry_msgs::Point eigen2point(Eigen::Vector3d vector);
-
 std::vector<double> eigen2std(const Eigen::Vector3d& v);
-
-geometry_msgs::Vector3 vectorNull();
-
-geometry_msgs::Vector3 vectorUniform(double a);
 
 int nChoosek(int n, int k);
 void linearTransformPoly(const Eigen::VectorXd& coeff_old, Eigen::VectorXd& coeff_new, double a, double b);
@@ -201,10 +135,6 @@ std::ostream& operator<<(std::ostream& out, const std::vector<T>& v)
   return out;
 }
 
-visualization_msgs::MarkerArray trajectory2ColoredMarkerArray(const mt::trajectory& data, double max_value, int increm,
-                                                              std::string ns, double scale, std::string color_type,
-                                                              int id_agent, int n_agents, double min_cost = 0.0,
-                                                              double max_cost = 0.0, double cost = 0.0,
-                                                              bool collides = false);
+std::string casadi_folder();
 
 #endif
