+++ panther_compression/compression/envs/MyEnvironment.py
@@ -3,8 +3,12 @@ import sys
 import numpy as np
 import copy
 from gym import spaces
-from compression.utils.other import ActionManager, ObservationManager, GTermManager, State, ObstaclesManager, getPANTHERparamsAsCppStruct, computeTotalTime, getObsAndGtermToCrossPath, posAccelYaw2TfMatrix
-from compression.utils.other import TfMatrix2RosQuatAndVector3, TfMatrix2RosPose
+from compression.utils.other import GTermManager, State, computeTotalTime, getObsAndGtermToCrossPath, posAccelYaw2TfMatrix
+from compression.utils.yaml_utils import getPANTHERparamsAsCppStruct
+from compression.utils.ActionManager import ActionManager
+from compression.utils.ObservationManager import ObservationManager
+from compression.utils.ObstaclesManager import ObstaclesManager
+from compression.utils.ros_utils import TfMatrix2RosQuatAndVector3, TfMatrix2RosPose
 from compression.utils.other import CostComputer
 from compression.utils.other import MyClampedUniformBSpline
 from compression.utils.other import listOf3dVectors2numpy3Xmatrix
