+++ panther_compression/compression/policies/StudentPolicy_noProbability.py
@@ -13,7 +13,9 @@ from stable_baselines3.common.torch_layers import (
     create_mlp,
 )
 
-from compression.utils.other import ActionManager, ObservationManager
+from compression.utils.ActionManager import ActionManager
+from compression.utils.ObservationManager import ObservationManager
+
 from colorama import init, Fore, Back, Style
