+++ panther_compression/compression/utils/other.py
@@ -11,6 +11,11 @@ import random
 import pytest
 import time
 
+from .ActionManager import ActionManager
+from .ObservationManager import ObservationManager
+from .ObstaclesManager import ObstaclesManager
+
+from .yaml_utils import getPANTHERparamsAsCppStruct
 
 import numpy.matlib
 
@@ -20,8 +25,7 @@ from stable_baselines3.common import utils
 ########
 
 #######
-import geometry_msgs.msg
-import tf.transformations 
+# import tf.transformations
 # from geometry_msgs.msg import PointStamped, TransformStamped, PoseStamped, Vector3, Quaternion, Pose
 #######
 
@@ -29,7 +33,6 @@ import tf.transformations
 from joblib import Parallel, delayed
 import multiprocessing
 ##
-
 
 class ExpertDidntSucceed(Exception):
 	  pass
@@ -106,43 +109,6 @@ def posAccelYaw2TfMatrix(w_pos, w_accel, yaw):
 	# print(w_T_b)
 
 	return TfMatrix(w_T_b)
-
-def getPANTHERparamsAsCppStruct():
-
-	params_yaml=readPANTHERparams();
-
-	params_yaml["b_T_c"]=np.array([[0, 0, 1, 0],
-								  [-1, 0, 0, 0],
-								  [0, -1, 0, 0],
-								  [0, 0, 0, 1]])
-
-	par=py_panther.parameters();
-
-	for key in params_yaml:
-		exec('%s = %s' % ('par.'+key, 'params_yaml["'+key+'"]')) #See https://stackoverflow.com/a/60487422/6057617 and https://www.pythonpool.com/python-string-to-variable-name/
-
-	return par
-
-
-def readPANTHERparams():
-
-	params_yaml_1=[];
-	with open(os.path.dirname(os.path.abspath(__file__)) + '/../../../panther/param/panther.yaml', "r") as stream:
-		try:
-			params_yaml_1=yaml.safe_load(stream)
-		except yaml.YAMLError as exc:
-			print(exc)
-
-	params_yaml_2=[];
-	with open(os.path.dirname(os.path.abspath(__file__)) + '/../../../panther/matlab/casadi_generated_files/params_casadi.yaml', "r") as stream:
-		try:
-			params_yaml_2=yaml.safe_load(stream)
-		except yaml.YAMLError as exc:
-			print(exc)
-	# params_yaml = dict(params_yaml_1.items() + params_yaml_2.items()) #Doesn't work in Python 3
-	params_yaml = {**params_yaml_1, **params_yaml_2}                        # NOTE: Python 3.5+ ONLY
-
-	return params_yaml;
 
 class Obstacle():
 	def __init__(self, ctrl_pts, bbox_inflated):
@@ -226,85 +192,6 @@ class GTermManager():
 
 	def get_w_GTermPos(self):
 		return self.w_gterm;
-
-
-class ObstaclesManager():
-
-    #The reason to create this here (instead of in the constructor) is that C++ objects created with pybind11 cannot be pickled by default (pickled is needed when parallelizing)
-    #See https://stackoverflow.com/a/68672/6057617
-    #Other option would be to do this: https://pybind11.readthedocs.io/en/stable/advanced/classes.html#pickling-support
-    #Note that pickle is needed when saving the Student Policy (which has a ObservationManager, which has a ObstaclesManager, which has a Fitter )
-	params=readPANTHERparams()
-	fitter = py_panther.Fitter(params["fitter_num_samples"]);
-
-	def __init__(self):
-		self.num_obs=1;
-		self.params=readPANTHERparams();
-		# self.fitter_total_time=params["fitter_total_time"];
-		self.fitter_num_seg=self.params["fitter_num_seg"];
-		self.fitter_deg_pos=self.params["fitter_deg_pos"];
-		self.fitter_total_time=self.params["fitter_total_time"];
-		self.fitter_num_samples=self.params["fitter_num_samples"];
-		# self.fitter = py_panther.Fitter(self.fitter_num_samples);
-		self.newRandomPos();
-
-	def newRandomPos(self):
-		self.random_pos=np.array([[random.uniform(-4.0, 4.0)],[random.uniform(-4.0, 4.0)],[random.uniform(1.0,1.0)]]);
-		self.random_offset=random.uniform(0.0, 10*math.pi)
-		self.random_scale=np.array([[random.uniform(0.5, 4.0)],[random.uniform(0.5, 4.0)],[random.uniform(0.5, 4.0)]]);
-		#self.random_pos=np.array([[2.5],[1.0],[1.0]]);
-
-	def setPos(self, pos):
-		self.random_pos=pos
-
-	def getNumObs(self):
-		return self.num_obs
-
-	def getCPsPerObstacle(self):
-		return self.fitter_num_seg + self.fitter_deg_pos
-
-	def getSizeAllObstacles(self):
-		#Size of the ctrl_pts + bbox
-		return self.num_obs*(3*self.getCPsPerObstacle() + 3) 
-
-	def getFutureWPosStaticObstacles(self):
-		w_obs=[];
-		for i in range(self.num_obs):
-			w_ctrl_pts_ob=np.array([[],[],[]]);
-			for j in range(self.fitter_num_seg + self.fitter_deg_pos):
-				w_ctrl_pts_ob=np.concatenate((w_ctrl_pts_ob, self.random_pos), axis=1)
-				# w_ctrl_pts_ob.append(np.array([[2],[2],[2]]))
-
-			# bbox_ob=np.array([[0.5],[0.5], [0.5]]);
-			bbox_inflated=np.array([[0.8],[0.8], [0.8]])+2*self.params["drone_radius"]*np.ones((3,1));
-			w_obs.append(Obstacle(w_ctrl_pts_ob, bbox_inflated))
-		return w_obs;
-
-	def getFutureWPosDynamicObstacles(self,t):
-		w_obs=[];
-		# trefoil=Trefoil(pos=self.random_pos, scale_x=1.0, scale_y=1.0, scale_z=1.0, offset=0.0, slower=1.5);
-		# novale=np.array([[4.0],[4.0],[1.0]]);
-		# print(f"Using offset={self.random_offset}")
-		# print(f"Using random_scale={self.random_scale}")
-
-		###HACK TO GENERATE A STATIC OBSTACLE
-		self.random_scale=np.zeros((3,1))
-		####################################
-
-		trefoil=Trefoil(pos=self.random_pos, scale=self.random_scale, offset=self.random_offset, slower=1.5);
-		for i in range(self.num_obs):
-
-			samples=[]
-			for t_interm in np.linspace(t, t + self.fitter_total_time, num=self.fitter_num_samples):#.tolist():
-				samples.append(trefoil.getPosT(t_interm))
-
-			w_ctrl_pts_ob_list=ObstaclesManager.fitter.fit(samples)
-
-			w_ctrl_pts_ob=listOf3dVectors2numpy3Xmatrix(w_ctrl_pts_ob_list)
-
-			bbox_inflated=np.array([[0.8],[0.8], [0.8]])+2*self.params["drone_radius"]*np.ones((3,1));
-			w_obs.append(Obstacle(w_ctrl_pts_ob, bbox_inflated))
-		return w_obs;
 
 class Trefoil():
 	def __init__(self, pos, scale, offset, slower):
@@ -494,233 +381,6 @@ def computeTotalTime(init_state, final_state, par_vmax, par_amax, par_factor_all
 def wrapInmPiPi(data): #https://stackoverflow.com/a/15927914
 	return (data + np.pi) % (2 * np.pi) - np.pi
 
-class ObservationManager():
-	def __init__(self):
-		self.obsm=ObstaclesManager();
-		#Observation =       [f_v, f_a, yaw_dot, f_g,  [f_ctrl_pts_o0], bbox_o0, [f_ctrl_pts_o1], bbox_o1 ,...]
-		#
-		# Where f_ctrl_pts_oi = [cp0.transpose(), cp1.transpose(), ...]
-		self.observation_size= 3 +  3 +   1    + 3   + self.obsm.getSizeAllObstacles();
-
-		params=readPANTHERparams();
-		self.v_max=np.array(params["v_max"]).reshape(3,1);
-		self.a_max=np.array(params["a_max"]).reshape(3,1);
-		self.j_max=np.array(params["j_max"]).reshape(3,1);
-		self.ydot_max=params["ydot_max"];
-		# self.max_dist2goal=params["max_dist2goal"];
-		self.max_dist2obs=params["max_dist2obs"];
-		self.max_side_bbox_obs=params["max_side_bbox_obs"];
-		self.Ra=params["Ra"]
-		ones13=np.ones((1,3));
-		#Note that the sqrt(3) is needed because the expert/student plan in f_frame --> bouding ball around the box v_max, a_max,... 
-		margin_v=math.sqrt(3) #math.sqrt(3)
-		margin_a=math.sqrt(3) #math.sqrt(3)
-		margin_ydot=1.5 
-		self.normalization_constant=np.concatenate((margin_v*self.v_max.T*ones13, margin_a*self.a_max.T*ones13, margin_ydot*self.ydot_max*np.ones((1,1)), self.Ra*ones13), axis=1)
-		for i in range(self.obsm.getNumObs()):
-			self.normalization_constant=np.concatenate((self.normalization_constant, self.max_dist2obs*np.ones((1,3*self.obsm.getCPsPerObstacle())), self.max_side_bbox_obs*ones13), axis=1)
-
-		# assert print("Shape observation=", observation.shape==)
-
-	def randomVel(self):
-		return np.random.uniform(-self.v_max,self.v_max)
-
-	def randomAccel(self):
-		return np.random.uniform(-self.a_max,self.a_max)
-
-	def randomYdot(self):
-		return np.random.uniform(-self.ydot_max,self.ydot_max, size=(1,1))
-
-	def randomYaw(self):
-		return wrapInmPiPi(np.random.uniform(-np.pi,np.pi, size=(1,1)))
-
-	def obsIsNormalized(self, observation_normalized):
-		# print(observation_normalized.shape)
-		assert observation_normalized.shape == self.getObservationShape()
-
-		return np.logical_and(observation_normalized >= -1, observation_normalized <= 1).all()
-
-	def assertObsIsNormalized(self, observation_normalized, msg_before=""):
-
-		if not self.obsIsNormalized(observation_normalized):
-			print(msg_before+"The observation is not normalized")
-			print(f"NORMALIZED VALUE={observation_normalized}")
-			observation=self.denormalizeObservation(observation_normalized)
-			print(f"VALUE={observation}")
-			self.printObservation(observation);
-			raise AssertionError()
-
-	def printObservation(self, obs):
-		print("----The observation is:")
-		print(Fore.BLUE +f"f_v.T={self.getf_v(obs).T}"+Style.RESET_ALL)
-		print(Fore.GREEN +f"f_a.T={self.getf_a(obs).T}"+Style.RESET_ALL)
-		print(Fore.MAGENTA +f"yaw_dot={self.getyaw_dot(obs)}"+Style.RESET_ALL)
-		print(Fore.CYAN +f"f_g.T={self.getf_g(obs).T}"+Style.RESET_ALL)
-
-		for i in range(self.obsm.getNumObs()):
-			print(f"Obstacle {i}:")
-			ctrl_pts=self.getCtrlPtsObstacleI(obs,i) 
-			bbox_inflated=self.getBboxInflatedObstacleI(obs,i)
-
-			print(Fore.YELLOW +f"ctrl_pts={ctrl_pts}"+Style.RESET_ALL)
-			print(Fore.BLUE +f"bbox_inflated.T={bbox_inflated.T}"+Style.RESET_ALL)
-
-		print("----------------------")
-
-	def get_f_StateFromf_obs(self,f_obs):
-		pos=np.zeros((3,1))
-		vel=self.getf_v(f_obs)
-		accel=self.getf_a(f_obs)
-		yaw=np.zeros((1,1))
-		yaw_dot=self.getyaw_dot(f_obs)
-		state=State(pos, vel, accel, yaw, yaw_dot)
-		return state
-
-
-	def getIndexStartObstacleI(self,i):
-		return 10+(self.obsm.getCPsPerObstacle()+3)*i
-
-	def getCtrlPtsObstacleI(self,obs,i):
-		index_start_obstacle_i=self.getIndexStartObstacleI(i)
-		ctrl_pts=[]; 
-		num_cps_per_obstacle=self.obsm.getCPsPerObstacle();
-		for j in range(num_cps_per_obstacle):
-			index_start_cpoint=index_start_obstacle_i+3*j
-			cpoint_j=obs[0,index_start_cpoint:index_start_cpoint+3].reshape(3,1)
-			ctrl_pts.append(cpoint_j)
-
-		return ctrl_pts
-
-	def getBboxInflatedObstacleI(self,obs,i):
-		index_start_obstacle_i=self.getIndexStartObstacleI(i)
-
-		tmp=index_start_obstacle_i+3*self.obsm.getCPsPerObstacle()
-
-		bbox_inflated=obs[0,tmp:tmp+4].reshape(3,1)
-
-		return bbox_inflated
-
-	def getf_v(self, obs):
-		return obs[0,0:3].reshape((3,1)) #Column vector
-
-	def getf_a(self, obs):
-		return obs[0,3:6].reshape((3,1)) #Column vector
-
-	def getyaw_dot(self, obs):
-		return obs[0,6].reshape((1,1)) 
-
-	def getf_g(self, obs):
-		return obs[0,7:10].reshape((3,1)) #Column vector
-
-	def getObstacles(self, obs):
-		# print("obs is= ", obs)
-
-		obstacles=[]
-		for i in range(self.obsm.getNumObs()):
-
-			ctrl_pts=self.getCtrlPtsObstacleI(obs,i) 
-			bbox_inflated=self.getBboxInflatedObstacleI(obs,i)
-
-			obstacle=py_panther.obstacleForOpt()
-
-			obstacle.ctrl_pts=ctrl_pts
-			obstacle.bbox_inflated=bbox_inflated
-
-			obstacles.append(obstacle)
-
-		return obstacles
-
-	def getInit_f_StateFromObservation(self, obs):
-		init_state=py_panther.state();  #Everything initialized as zero
-		init_state.pos= np.array([[0.0],[0.0],[0.0]]);#Because it's in f frame
-		init_state.vel= self.getf_v(obs);
-		init_state.accel= self.getf_a(obs);
-		init_state.yaw= 0.0  #Because it's in f frame
-		init_state.dyaw = self.getyaw_dot(obs);
-		return init_state
-
-	def getFinal_f_StateFromObservation(self, obs):
-		final_state=py_panther.state();  #Everything initialized as zero
-		final_state.pos= self.getf_g(obs);
-		# final_state.vel= 
-		# final_state.accel= 
-		# final_state.yaw= 
-		# final_state.dyaw = 
-		return final_state
-
-	def getNanObservation(self):
-		return np.full(self.getObservationShape(), np.nan)
-
-	def isNanObservation(self, obs):
-		return np.isnan(np.sum(obs))
-
-	#Normalize in [-1,1]
-	def normalizeObservation(self, observation):
-		# print("Shape observation=", observation.shape)
-		# print("Shape normalization_constant=", self.normalization_constant.shape)
-		# print("obsm.getSizeAllObstacles()=", self.obsm.getSizeAllObstacles())
-
-		observation_normalized=observation/self.normalization_constant;
-		# assertIsNormalized(observation_normalized)
-		# assert np.logical_and(observation_normalized >= -1, observation_normalized <= 1).all()
-		return observation_normalized;
-
-	def denormalizeObservation(self,observation_normalized):
-		# assert np.logical_and(observation_normalized >= -1, observation_normalized <= 1).all()
-
-		# assertIsNormalized(observation_normalized)
-		# assert np.logical_and(observation_normalized >= -1, observation_normalized <= 1).all(), f"observation_normalized= {observation_normalized}" 
-		observation=observation_normalized*self.normalization_constant;
-		return observation;
-
-	# def denormalize(self, )
-
-	def getObservationSize(self):
-		return self.observation_size
-
-	def getObservationShape(self):
-		return (1,self.observation_size)
-
-	def getRandomObservation(self):
-		random_observation=self.denormalizeObservation(self.getRandomNormalizedObservation())
-		return random_observation
-
-	def getRandomNormalizedObservation(self):
-		random_normalized_observation=np.random.uniform(-1,1, size=self.getObservationShape())
-		return random_normalized_observation
-
-	def get_fObservationFrom_w_stateAnd_w_gtermAnd_w_obstacles(self,w_state, w_gterm_pos, w_obstacles):
-
-		f_gterm_pos=w_state.f_T_w * w_gterm_pos
-
-		dist2gterm=np.linalg.norm(f_gterm_pos);
-		f_g= min(dist2gterm-1e-4, self.Ra)*normalize(f_gterm_pos)
-		#f_g= self.Ra*normalize(f_gterm_pos)
-		# print("w_state.f_vel().flatten()= ", w_state.f_vel().flatten())
-		# print("w_state.f_accel().flatten()= ", w_state.f_accel().flatten())
-		# print("w_state.f_accel().flatten()= ", w_state.f_accel().flatten())
-		observation=np.concatenate((w_state.f_vel().flatten(), w_state.f_accel().flatten(), w_state.yaw_dot.flatten(), f_g.flatten()));
-
-		#Convert obs to f frame and append ethem to observation
-		for w_obstacle in w_obstacles:
-			assert type(w_obstacle.ctrl_pts).__module__ == np.__name__, "the ctrl_pts should be a numpy matrix, not a list"
-			observation=np.concatenate((observation, (w_state.f_T_w*w_obstacle.ctrl_pts).flatten(order='F'), (w_obstacle.bbox_inflated).flatten()))
-
-
-		observation=observation.reshape(self.getObservationShape())
-
-		# print("observation= ", observation)
-
-		assert observation.shape == self.getObservationShape()
-
-		return observation;
-
-	def getNormalized_fObservationFrom_w_stateAnd_w_gtermAnd_w_obstacles(self, w_state, w_gterm_pos, w_obstacles):
-		f_observation=self.get_fObservationFrom_w_stateAnd_w_gtermAnd_w_obstacles(w_state, w_gterm_pos, w_obstacles)
-		f_observationn=self.normalizeObservation(f_observation) #Observation normalized
-		return f_observationn
-
-
 def numpy3XmatrixToListOf3dVectors(data):
 	data_list=[]
 	for i in range(data.shape[1]):
@@ -743,269 +403,6 @@ def getZeroState():
 
 	return zero_state
 
-class ActionManager():
-	def __init__(self):
-		params=readPANTHERparams();
-		self.deg_pos=params["deg_pos"];
-		self.deg_yaw=params["deg_yaw"];
-		self.num_seg=params["num_seg"];
-		self.use_closed_form_yaw_student=params["use_closed_form_yaw_student"];
-
-		# Define action and observation space
-
-		# action = np.array([ctrl_points_pos   ctrl_points_yaw  total time, prob that traj])
-		#
-		# --> where ctrlpoints_pos has the ctrlpoints that are not determined by init pos/vel/accel, and final vel/accel
-		#     i.e., len(ctrlpoints_pos)= (num_seg_pos + deg_pos - 1 + 1) - 3 - 2 = num_seg_pos + deg_pos - 5;
-		#	  and ctrl_points_pos=[cp3.transpose(), cp4.transpose(),...]
-		#
-		# --> where ctrlpoints_yaw has the ctrlpoints that are not determined by init pos/vel, and final vel
-		#     i.e., len(ctrlpoints_yaw)= (num_seg_yaw + deg_yaw - 1 + 1) - 2 - 1 = num_seg_yaw + deg_yaw - 3;
-
-		self.num_traj_per_action=params["num_of_trajs_per_replan"];
-
-		self.total_num_pos_ctrl_pts = self.num_seg + self.deg_pos
-		self.traj_size_pos_ctrl_pts = 3*(self.total_num_pos_ctrl_pts - 5);
-		self.traj_size_yaw_ctrl_pts = (self.num_seg + self.deg_yaw - 3);
-		# self.traj_size = self.traj_size_pos_ctrl_pts + self.traj_size_yaw_ctrl_pts + 1 + 1; # Last two numbers are time and prob that traj is real
-		self.traj_size = self.traj_size_pos_ctrl_pts + self.traj_size_yaw_ctrl_pts + 1; # Last number is time
-		self.action_size = self.num_traj_per_action*self.traj_size;
-		self.Npos = self.num_seg + self.deg_pos-1;
-
-		self.max_dist2BSPoscPoint=params["max_dist2BSPoscPoint"];
-		self.max_yawcPoint=4e3*math.pi;
-		self.fitter_total_time=params["fitter_total_time"];
-
-		# print("self.max_dist2BSPoscPoint= ", self.max_dist2BSPoscPoint)
-		# print("self.max_yawcPoint= ", self.max_yawcPoint)
-		self.normalization_constant_traj=np.concatenate((self.max_dist2BSPoscPoint*np.ones((1, self.traj_size_pos_ctrl_pts)), \
-													self.max_yawcPoint*np.ones((1, self.traj_size_yaw_ctrl_pts))), axis=1)
-
-		self.normalization_constant=np.matlib.repmat(self.normalization_constant_traj, self.num_traj_per_action, 1)
-
-
-	def actionIsNormalized(self, action_normalized):
-		assert action_normalized.shape == self.getActionShape()
-
-		return np.logical_and(action_normalized >= -1, action_normalized <= 1).all()
-
-	def assertActionIsNormalized(self, action_normalized, msg_before=""):
-		if not self.actionIsNormalized(action_normalized):
-			print(msg_before+"The observation is not normalized")
-			print(f"NORMALIZED VALUE={action_normalized}")
-			action=self.denormalizeAction(action_normalized)
-			print(f"VALUE={action}")
-			# self.printObservation(observation);
-			raise AssertionError()
-
-	def assertAction(self,action):
-		assert action.shape==self.getActionShape(), f"[Env] ERROR: action.shape={action.shape} but should be={self.getActionShape()}"
-		assert not np.isnan(np.sum(action)), f"Action has nan"
-
-
-	def getDummyOptimalNormalizedAction(self):
-		action=self.getDummyOptimalAction();
-		return self.normalizeAction(action)
-
-	def getDummyOptimalAction(self):
-		# return 0.6*np.ones(self.getActionShape())
-
-		dummy=np.ones((self.num_traj_per_action,self.traj_size))
-
-		for i in range((dummy.shape[0])):
-			for j in range(0,self.traj_size_pos_ctrl_pts,3):
-				dummy[i,j]=i+j/10
-
-		return dummy
-
-
-	def getNanAction(self):
-		return np.full(self.getActionShape(), np.nan)
-
-	def isNanAction(self, act):
-		return np.isnan(np.sum(act))
-
-	def normalizeAction(self, action):
-		action_normalized=np.empty(action.shape)
-		action_normalized[:,0:-1]=action[:,0:-1]/self.normalization_constant #Elementwise division
-		action_normalized[:,-1]=(2.0/self.fitter_total_time)*action[:,-1]-1 #Note that action[0,-1] is in [0, fitter_total_time]
-		# action_normalized[:,-1]=(2.0/1.0)*action[:,-1]-1 #Note that action[0,-1] is in [0, 1]
-
-		for index_traj in range(self.num_traj_per_action):
-			time_normalized=self.getTotalTime(action_normalized, index_traj);
-			slack=1-abs(time_normalized);
-			if(slack<0):
-				if abs(slack)<1e-5: #Can happen due to the tolerances in the optimization
-					# print(f"Before= {action_normalized[0,-1]}")
-					action_normalized[index_traj,-1]=np.clip(time_normalized, -1.0, 1.0) #Saturate within limits
-					# print(f"After= {action_normalized[0,-1]}")
-				else:
-					assert False, f"time_normalized={time_normalized}"
-
-
-		# assert np.logical_and(action_normalized >= -1, action_normalized <= 1).all(), f"action_normalized={action_normalized}, last element={action_normalized[0,-1]}"
-		return action_normalized;
-
-	# def getProb(self,action, index_traj):
-	# 	return action[index_traj,-1]
-
-	def getTotalTime(self,action, index_traj):
-		return action[index_traj,-1]
-
-	def getPosCtrlPts(self, action, index_traj):
-		return action[index_traj,0:self.traj_size_pos_ctrl_pts].reshape((3,-1), order='F')
-
-	def getYawCtrlPts(self, action, index_traj):
-		return action[index_traj,self.traj_size_pos_ctrl_pts:-1].reshape((1,-1));
-
-	# def getProbTraj(self,traj):
-	# 	return self.getProb(traj, 0)
-
-	def getTotalTimeTraj(self,traj):
-		return self.getTotalTime(traj, 0)
-
-	def getPosCtrlPtsTraj(self, traj):
-		return self.getPosCtrlPts(traj,0)
-
-	def getYawCtrlPtsTraj(self, traj):
-		return self.getYawCtrlPts(traj, 0);
-
-	def printAction(self, action):
-		print(f"Raw={action}")
-		for index_traj in range(self.num_traj_per_action):
-			print(Fore.WHITE+f"Traj {index_traj}: "+Style.RESET_ALL)
-			print(Fore.RED+f"  f pos ctrl_pts=\n{self.getPosCtrlPts(action, index_traj)}"+Style.RESET_ALL)
-			print(Fore.YELLOW+f"  f yaw ctrl_pts={self.getYawCtrlPts(action, index_traj)}"+Style.RESET_ALL)
-			print(Fore.BLUE+f"  Total time={self.getTotalTime(action, index_traj)}"+Style.RESET_ALL)
-
-	def denormalizeAction(self, action_normalized):
-		# assert np.logical_and(action_normalized >= -1, action_normalized <= 1).all(), f"action_normalized={action_normalized}"
-		action=np.empty(action_normalized.shape)
-		action[:,0:-1]=action_normalized[:,0:-1]*self.normalization_constant #Elementwise multiplication
-		action[:,-1]=(self.fitter_total_time/2.0)*(action_normalized[:,-1]+1) #Note that action[:,-2] is in [0, fitter_total_time]
-		# action[:,-1]=(1.0/2.0)*(action_normalized[:,-1]+1) #Note that action[:,-1] is in [0, 1]
-		return action
-
-	def denormalizeTraj(self, traj_normalized):
-		#TODO: this could be done more efficiently
-		dummy_action_normalized=np.matlib.repmat(traj_normalized, self.num_traj_per_action, 1)
-		dummy_action=self.denormalizeAction(dummy_action_normalized)
-		traj=dummy_action[0,:].reshape(traj_normalized.shape)
-		return traj
-
-	def getActionSize(self):
-		return self.action_size
-
-	def getActionShape(self):
-		return (self.num_traj_per_action,self.traj_size)
-
-	def getTrajShape(self):
-		return (1,self.traj_size)
-
-	def getRandomAction(self):
-		random_action=self.denormalizeAction(self.getRandomNormalizedAction())
-		return random_action
-
-	def getRandomNormalizedAction(self):
-		random_normalized_action=np.random.uniform(-1,1, size=self.getActionShape())
-		return random_normalized_action
-
-	def getDegPos(self):
-		return self.deg_pos;
-
-	def getDegYaw(self):
-		return self.deg_yaw;
-
-	def f_trajAnd_w_State2_w_pos_ctrl_pts_and_knots(self, f_traj, w_state):
-
-		assert type(w_state)==State
-
-		total_time=self.getTotalTimeTraj(f_traj)
-
-		knots_pos=generateKnotsForClampedUniformBspline(0.0, total_time, self.deg_pos, self.num_seg)
-
-		f_pos_ctrl_pts = self.getPosCtrlPtsTraj(f_traj)
-
-		#Convert to w frame
-		w_pos_ctrl_pts = w_state.w_T_f * f_pos_ctrl_pts;
-
-		pf=w_pos_ctrl_pts[:,-1].reshape((3,1)); #Assumming final vel and accel=0 
-		p0=w_state.w_pos
-		v0=w_state.w_vel
-		a0=w_state.w_accel
-
-		p=self.deg_pos;
-
-		t1 = knots_pos[1];
-		t2 = knots_pos[2];
-		tpP1 = knots_pos[p + 1];
-		t1PpP1 = knots_pos[1 + p + 1];
-
-		
-		# // See Mathematica Notebook
-		q0_pos = p0;
-		q1_pos = p0 + (-t1 + tpP1) * v0 / p;
-		q2_pos = (p * p * q1_pos - (t1PpP1 - t2) * (a0 * (t2 - tpP1) + v0) - p * (q1_pos + (-t1PpP1 + t2) * v0)) / ((-1 + p) * p);
-
-		w_pos_ctrl_pts=np.concatenate((q0_pos, q1_pos, q2_pos, w_pos_ctrl_pts, pf, pf), axis=1) #Assumming final vel and accel=0
-
-		return w_pos_ctrl_pts, knots_pos
-
-
-	def f_trajAnd_w_State2_w_yaw_ctrl_pts_and_knots(self, f_traj, w_state):
-		total_time=self.getTotalTimeTraj(f_traj)
-
-		knots_yaw=generateKnotsForClampedUniformBspline(0.0, total_time, self.deg_yaw, self.num_seg)
-
-		f_yaw_ctrl_pts= self.getYawCtrlPtsTraj(f_traj)
-
-		w_yaw_ctrl_pts =  f_yaw_ctrl_pts + w_state.w_yaw*np.ones(f_yaw_ctrl_pts.shape);
-
-
-		y0=w_state.w_yaw
-		y_dot0=w_state.yaw_dot
-		yf=w_yaw_ctrl_pts[0,-1].reshape((1,1)); #Assumming final vel =0
-
-		p=self.deg_yaw;
-		t1 = knots_yaw[1];
-		tpP1 = knots_yaw[p + 1];
-		q0_yaw = y0;
-		q1_yaw = y0 + (-t1 + tpP1) * y_dot0 / p;
-
-		w_yaw_ctrl_pts=np.concatenate((q0_yaw, q1_yaw, w_yaw_ctrl_pts, yf), axis=1) #Assumming final vel and accel=0
-
-		return w_yaw_ctrl_pts, knots_yaw
-
-	def f_trajAnd_w_State2w_ppSolOrGuess(self, f_traj, w_state): #pp stands for py_panther
-		w_p_ctrl_pts,knots_p=self.f_trajAnd_w_State2_w_pos_ctrl_pts_and_knots(f_traj, w_state)
-		w_y_ctrl_pts,knots_y=self.f_trajAnd_w_State2_w_yaw_ctrl_pts_and_knots(f_traj, w_state)
-
-		#Create and fill solOrGuess object
-		w_sol_or_guess=py_panther.solOrGuess();
-
-		w_sol_or_guess.qp=numpy3XmatrixToListOf3dVectors(w_p_ctrl_pts)
-		w_sol_or_guess.qy=numpy3XmatrixToListOf3dVectors(w_y_ctrl_pts)
-
-		w_sol_or_guess.knots_p=knots_p
-		w_sol_or_guess.knots_y=knots_y
-
-		w_sol_or_guess.solver_succeeded=True
-		# w_sol_or_guess.prob=self.getProbTraj(f_traj);
-		w_sol_or_guess.cost=0.0 #TODO
-		w_sol_or_guess.obst_avoidance_violation=0.0 #TODO
-		w_sol_or_guess.dyn_lim_violation=0.0 #TODO
-		w_sol_or_guess.is_guess=False #TODO
-
-		w_sol_or_guess.deg_p=self.deg_pos #TODO
-		w_sol_or_guess.deg_y=self.deg_yaw #TODO
-
-		return w_sol_or_guess
-
-	def f_trajAnd_f_State2f_ppSolOrGuess(self, f_traj, f_state):
-		assert np.linalg.norm(f_state.w_pos)<1e-7, "The pos should be zero"
-		assert np.linalg.norm(f_state.w_yaw)<1e-7, "The yaw should be zero"
-		return self.f_trajAnd_w_State2w_ppSolOrGuess(f_traj, f_state)
-
 # 	def f_obs_f_traj_2f_ppSolOrGuess(self, f_traj): #pp stands for py_panther
 # 		zero_state=State(np.zeros((3,1)), np.zeros((3,1)), np.zeros((3,1)), np.zeros((1,1)), np.zeros((1,1)))
 # class State():
@@ -1013,69 +410,6 @@ class ActionManager():
 # 		pos=np.zeros(3,1)
 # 		vel=np.zeros(3,1)
 # 		return self.f_trajAnd_w_State2w_ppSolOrGuess(f_traj, zero_state)
-
-
-	def getTrajFromAction(self, action, index_traj):
-		return action[index_traj,:].reshape(1,-1)
-
-	def f_trajAnd_w_State2wBS(self, f_traj, w_state, no_deriv=False):
-
-
-		w_pos_ctrl_pts,_=self.f_trajAnd_w_State2_w_pos_ctrl_pts_and_knots(f_traj, w_state)
-		w_yaw_ctrl_pts,_=self.f_trajAnd_w_State2_w_yaw_ctrl_pts_and_knots(f_traj, w_state)
-		total_time=self.getTotalTimeTraj(f_traj)
-
-
-		# print(f"f_action={f_action}")
-		# print(f"total_time={total_time}")
-		w_posBS = MyClampedUniformBSpline(0.0, total_time, self.deg_pos, 3, self.num_seg, w_pos_ctrl_pts, no_deriv) #def __init__():[BSpline(knots_pos, w_pos_ctrl_pts[0,:], self.deg_pos)
-
-		w_yawBS = MyClampedUniformBSpline(0.0, total_time, self.deg_yaw, 1, self.num_seg, w_yaw_ctrl_pts, no_deriv)
-		
-		return w_posBS, w_yawBS
-
-	def f_trajAnd_f_State2fBS(self, f_traj, f_state, no_deriv=False):
-		assert np.linalg.norm(f_state.w_pos)<1e-7, "The pos should be zero"
-		assert np.linalg.norm(f_state.w_yaw)<1e-7, "The yaw should be zero"
-		f_posBS, f_yawBS = self.f_trajAnd_w_State2wBS(f_traj, f_state, no_deriv)
-		return f_posBS, f_yawBS
-
-	def solOrGuess2traj(self, sol_or_guess):
-		traj=np.array([[]]);
-
-		#Append position control points
-		for i in range(3,len(sol_or_guess.qp)-2):
-			# print(sol_or_guess.qp[i].reshape(1,3))
-			traj=np.concatenate((traj, sol_or_guess.qp[i].reshape(1,3)), axis=1)
-
-		#Append yaw control points
-		for i in range(2,len(sol_or_guess.qy)-1):
-			traj=np.concatenate((traj, np.array([[sol_or_guess.qy[i]]])), axis=1)
-
-		#Append total time
-		assert sol_or_guess.knots_p[-1]==sol_or_guess.knots_y[-1]
-		assert sol_or_guess.knots_p[0]==0
-		assert sol_or_guess.knots_y[0]==0
-		traj=np.concatenate((traj, np.array([[sol_or_guess.knots_p[-1]]])), axis=1)
-
-		#Append prob of that traj
-		# traj=np.concatenate((traj, np.array( [[sol_or_guess.prob]]  )), axis=1)
-
-
-		assert traj.shape==self.getTrajShape()
-
-		return traj
-
-	def solsOrGuesses2action(self, sols_or_guesses):
-		assert len(sols_or_guesses)==self.num_traj_per_action
-		action=np.empty((0,self.traj_size));
-		for sol_or_guess in sols_or_guesses:
-			traj=self.solOrGuess2traj(sol_or_guess)
-			action=np.concatenate((action, traj), axis=0)
-
-		assert action.shape==self.getActionShape()
-		return action
-
 
 class ClosedFormYawSubstituter():
 	def __init__(self):
@@ -1185,39 +519,6 @@ class StudentCaller():
 
 	def getIndexBestTraj(self):
 		return self.costs_and_violations_of_action.index_best_traj
-
-
-def TfMatrix2RosQuatAndVector3(tf_matrix):
-
-  translation_ros=geometry_msgs.msg.Vector3();
-  rotation_ros=geometry_msgs.msg.Quaternion();
-
-  translation=tf_matrix.translation();
-  translation_ros.x=translation[0];
-  translation_ros.y=translation[1];
-  translation_ros.z=translation[2];
-  # q=tr.quaternion_from_matrix(w_state.w_T_f.T)
-  quaternion=tf.transformations.quaternion_from_matrix(tf_matrix.T) #See https://github.com/ros/geometry/issues/64
-  rotation_ros.x=quaternion[0] #See order at http://docs.ros.org/en/jade/api/tf/html/python/transformations.html
-  rotation_ros.y=quaternion[1]
-  rotation_ros.z=quaternion[2]
-  rotation_ros.w=quaternion[3]
-
-  return rotation_ros, translation_ros
-
-def TfMatrix2RosPose(tf_matrix):
-
-  rotation_ros, translation_ros=TfMatrix2RosQuatAndVector3(tf_matrix);
-
-  pose_ros=geometry_msgs.msg.Pose();
-  pose_ros.position.x=translation_ros.x
-  pose_ros.position.y=translation_ros.y
-  pose_ros.position.z=translation_ros.z
-
-  pose_ros.orientation=rotation_ros
-
-  return pose_ros
-
 
 class CostsAndViolationsOfAction():
 	def __init__(self, costs, obst_avoidance_violations, dyn_lim_violations, index_best_traj):
