+++ panther_compression/test_infeasibility_ipopt.py
@@ -2,8 +2,11 @@ import sys
 import numpy as np
 import copy
 from random import random, shuffle
-from compression.utils.other import ActionManager, ObservationManager, getPANTHERparamsAsCppStruct, ExpertDidntSucceed, computeTotalTime
+from compression.utils.other import ExpertDidntSucceed, computeTotalTime
+from compression.utils.yaml_utils import getPANTHERparamsAsCppStruct
 from compression.utils.other import numpy3XmatrixToListOf3dVectors, computeTotalTime
+from compression.utils.ActionManager import ActionManager
+from compression.utils.ObservationManager import ObservationManager
 from colorama import init, Fore, Back, Style
 import py_panther
 import math
