+++ panther_compression/tests.py
@@ -1,4 +1,6 @@
-from compression.utils.other import ActionManager, posAccelYaw2TfMatrix, State, ObservationManager
+from compression.utils.other import posAccelYaw2TfMatrix, State
+from compression.utils.ActionManager import ActionManager
+from compression.utils.ObservationManager import ObservationManager
 import numpy as np
 
 np.set_printoptions(linewidth=np.nan)
